requirements.txt
streamlit
pandas
plotly
reportlab
google-auth-oauthlib
google-api-python-client
pdfplumber
requests
